Add Energy
Added player energy and action energy, visuallized by a energybar above the player actions. Whenever a player choses an action, some energy will be removed. If the player doesn't have enough energy for an action, it will be locked. On the start of the player's next turn, all energy will be refilled.